+++ data/src/main/java/com/example/nguyenthanhtungh/data/di/RepositoryModule.kt
@@ -9,27 +9,9 @@ import com.example.nguyenthanhtungh.domain.repository.PhotoRepository
 import org.koin.dsl.module.module
 
 val repositoryModule = module(override = true) {
-    single { createUrlsEntityMapper() }
-    single { createUserEntityMapper() }
-    single { createCoverPhotoEntityMapper(get()) }
-    single { createCollectionEntityMapper(get()) }
-    single { createPhotoEntityMapper(get(), get()) }
-
     single { createCollectionRepository(get(), get()) }
     single { createPhotoRepository(get(), get()) }
 }
-
-fun createUrlsEntityMapper() = UrlsEntityMapper()
-
-fun createUserEntityMapper() = UserEntityMapper()
-
-fun createCoverPhotoEntityMapper(urlsEntityMapper: UrlsEntityMapper) = CoverPhotoEntityMapper(urlsEntityMapper)
-
-fun createCollectionEntityMapper(coverPhotoEntityMapper: CoverPhotoEntityMapper) =
-    CollectionEntityMapper(coverPhotoEntityMapper)
-
-fun createPhotoEntityMapper(userEntityMapper: UserEntityMapper, urlsEntityMapper: UrlsEntityMapper) =
-    PhotoEntityMapper(userEntityMapper, urlsEntityMapper)
 
 fun createCollectionRepository(
     apiService: ApiService,
