+++ data/src/main/java/com/example/nguyenthanhtungh/data/source/repository/PhotoRepositoryImpl.kt
@@ -7,12 +7,12 @@ import com.example.nguyenthanhtungh.domain.repository.PhotoRepository
 import io.reactivex.Single
 
 class PhotoRepositoryImpl(
-    val apiService: ApiService,
+    private val apiService: ApiService,
     private val photoEntityMapper: PhotoEntityMapper
 ) : PhotoRepository {
     override fun getListPhoto(id: String): Single<List<Photo>> {
-        return apiService.getCollectionPhotos(id).map {
-            it.map { photoEntityMapper.mapToDomain(it) }
+        return apiService.getCollectionPhotos(id).map { listPhotos ->
+            listPhotos.map { photoEntityMapper.mapToDomain(it) }
         }
     }
 }
