+++ domain/src/main/java/com/example/nguyenthanhtungh/domain/model/Collection.kt
@@ -3,7 +3,7 @@ package com.example.nguyenthanhtungh.domain.model
 import com.example.nguyenthanhtungh.domain.base.BaseModel
 
 class Collection(
-    val id: String? = null,
+    val id: String = "",
     val title: String? = null,
     val totalPhoto: Int? = null,
     val coverPhoto: CoverPhoto? = null
