+++ domain/src/main/java/com/example/nguyenthanhtungh/domain/model/Photo.kt
@@ -3,7 +3,7 @@ package com.example.nguyenthanhtungh.domain.model
 import com.example.nguyenthanhtungh.domain.base.BaseModel
 
 class Photo(
-    val id: String? = null,
+    val id: String = "",
     val description: String? = null,
     val user: User? = null,
     val urls: Urls? = null,
